requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.12.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Already in pyproject.toml (optional for app)
yt-dlp>=2023.12.0
youtube-transcript-api>=0.6.0